# openapi.yml
openapi: 3.0.3
info:
  title: Hospital Management System API
  description: |
    A comprehensive web-based Hospital Management System built with Flask.

    This API provides role-based access control for three user types:
    - **Admin**: Manages doctors, patients, and appointments
    - **Doctor**: Manages appointments, patient records, and availability
    - **Patient**: Books appointments and accesses treatment history

    ## Authentication
    The API uses session-based authentication with Flask-Login.
    Login is required for most endpoints with role-based access control.

    ## Security
    - Bcrypt password hashing
    - CSRF protection on all forms
    - SQL injection prevention via ORM
    - Role-based access control (RBAC)
  version: 0.5.0
  contact:
    name: IIT Madras Student
    email: admin@hospital.com
  license:
    name: Educational Use Only
    url: https://www.iitm.ac.in
servers:
  - url: http://127.0.0.1:5000
    description: Development server
  - url: http://localhost:5000
    description: Local server
...
paths:
...
  /patient/dashboard:
    get:
      tags:
        - Patient
      summary: Patient dashboard
      description: Display patient dashboard with upcoming appointments
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Dashboard displayed
        '403':
          description: Access denied - Patient role required
  /patient/doctors:
    get:
      tags:
        - Patient
      summary: Search doctors
      description: Search and filter doctors by specialization
      security:
        - sessionAuth: []
      parameters:
        - name: specialization
          in: query
          description: Filter by specialization
          schema:
            type: string
        - name: search
          in: query
          description: Search by doctor name
          schema:
            type: string
      responses:
        '200':
          description: List of doctors
  /patient/doctors/{id}:
    get:
      tags:
        - Patient
      summary: Doctor details
      description: View doctor profile and availability
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Doctor details displayed
        '404':
          description: Doctor not found
  /patient/appointments/book/{doctor_id}:
    get:
      tags:
        - Patient
      summary: Book appointment form
      description: Display appointment booking form with available slots
      security:
        - sessionAuth: []
      parameters:
        - name: doctor_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Booking form displayed
        '404':
          description: Doctor not found
    post:
      tags:
        - Patient
      summary: Create appointment
      description: Book an appointment with selected doctor
      security:
        - sessionAuth: []
      parameters:
        - name: doctor_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AppointmentCreate'
      responses:
        '302':
          description: Appointment booked successfully
        '400':
          description: Validation error or time slot conflict
        '404':
          description: Doctor not found
  /patient/appointments:
    get:
      tags:
        - Patient
      summary: My appointments
      description: View all patient appointments
      security:
        - sessionAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [Booked, Completed, Cancelled]
      responses:
        '200':
          description: List of appointments
...
  /patient/treatment-history:
    get:
      tags:
        - Patient
      summary: Treatment history
      description: View complete treatment history with diagnoses and prescriptions
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Treatment history displayed
  /patient/profile:
    get:
      tags:
        - Patient
      summary: View profile
      description: Display patient profile
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Profile displayed
    post:
      tags:
        - Patient
      summary: Update profile
      description: Update patient profile information
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatientProfileUpdate'
      responses:
        '302':
          description: Profile updated successfully

# ==================== COMPONENTS ====================
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Flask session-based authentication
  schemas:
...
    AppointmentCreate:
      type: object
      required:
        - appointment_date
        - appointment_time
      properties:
        appointment_date:
          type: string
          format: date
          example: '2025-12-01'
        appointment_time:
          type: string
          example: '10:00 AM'
        reason:
          type: string
          example: 'Regular checkup'
...
    PatientProfileUpdate:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        age:
          type: integer
          example: 36
        gender:
          type: string
          enum: [Male, Female, Other]
        phone:
          type: string
          example: '+91 9876543210'
        address:
          type: string
          example: '123 Main St, Mumbai'

    # Response Models
...
security:
  - sessionAuth: []